+++ lib/crimson.dart
@@ -3,3 +3,4 @@ library crimson;
 export 'src/annotations.dart' hide JsonIgnore, JsonKebabCase, JsonSnakeCase;
 export 'src/crimson.dart';
 export 'src/crimson_writer.dart';
+export 'src/json_type.dart';