+++ lib/src/json_type.dart
@@ -1,0 +1,20 @@
+/// The type of a JSON value.
+enum JsonType {
+  /// A JSON object.
+  object,
+
+  /// A JSON array.
+  array,
+
+  /// A String.
+  string,
+
+  /// An integer or double.
+  number,
+
+  /// True or false.
+  bool,
+
+  /// Null.
+  nil
+}